# pytorrent/openapi/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "error": {
            "type": "string"
          },
          "ok": {
            "enum": [
              false
            ],
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "error"
        ],
        "type": "object"
      },
      "ApiOk": {
        "properties": {
          "ok": {
            "enum": [
              true
            ],
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "type": "object"
      },
...
      "AutomationExportResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiOk"
          },
          {
            "properties": {
              "count": {
                "type": "integer"
              },
              "export": {
                "additionalProperties": true,
                "type": "object"
              }
            },
            "type": "object"
          }
        ]
      },
      "AutomationImportResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiOk"
          },
          {
            "properties": {
              "imported": {
                "type": "integer"
              },
              "rules": {
                "items": {
                  "$ref": "#/components/schemas/AutomationRule"
                },
                "type": "array"
              }
            },
            "type": "object"
          }
        ]
      },
      "AutomationListResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiOk"
          },
          {
            "properties": {
              "history": {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              "rules": {
                "items": {
                  "$ref": "#/components/schemas/AutomationRule"
                },
                "type": "array"
              }
            },
            "type": "object"
          }
        ]
      },
      "AutomationRule": {
        "additionalProperties": true,
        "type": "object"
      },
      "AutomationRunResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiOk"
          },
          {
            "properties": {
              "history": {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              "result": {
                "additionalProperties": true,
                "type": "object"
              },
              "rules": {
                "items": {
                  "$ref": "#/components/schemas/AutomationRule"
                },
                "type": "array"
              }
            },
            "type": "object"
          }
        ]
      },
...
      "CleanupCacheSummary": {
        "properties": {
          "profile_id": {
            "type": "integer"
          },
          "profile_rows": {
            "type": "integer"
          },
          "runtime_items": {
            "type": "integer"
          },
          "torrent_stats_rows": {
            "type": "integer"
          },
          "tracker_rows": {
            "type": "integer"
          }
        },
        "required": [
          "profile_id",
          "profile_rows",
          "runtime_items"
        ],
        "type": "object"
      },
...
      "CleanupSummary": {
        "properties": {
          "automation_history_total": {
            "type": "integer"
          },
          "cache": {
            "$ref": "#/components/schemas/CleanupCacheSummary"
          },
          "database": {
            "properties": {
              "error": {
                "type": "string"
              },
              "path": {
                "type": "string"
              },
              "size": {
                "format": "int64",
                "type": "integer"
              },
              "size_h": {
                "type": "string"
              }
            },
            "type": "object"
          },
          "jobs_clearable": {
            "type": "integer"
          },
          "jobs_total": {
            "type": "integer"
          },
          "operation_logs_total": {
            "type": "integer"
          },
          "planner_history_total": {
            "type": "integer"
          },
          "retention_days": {
            "properties": {
              "automation_history": {
                "type": "integer"
              },
              "jobs": {
                "type": "integer"
              },
              "operation_logs": {
                "type": "integer"
              },
              "planner_history": {
                "type": "integer"
              },
              "smart_queue_history": {
                "type": "integer"
              }
            },
            "type": "object"
          },
          "smart_queue_history_total": {
            "type": "integer"
          }
        },
        "required": [
          "jobs_total",
          "jobs_clearable",
          "smart_queue_history_total",
          "retention_days",
          "database",
          "cache",
          "planner_history_total",
          "operation_logs_total"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "apiKeyAuth": {
        "description": "Optional per-user API token. Tokens are accepted only when PYTORRENT_AUTH_ENABLE=1. When auth is disabled, API endpoints do not require credentials.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "bearerApiToken": {
        "description": "Optional per-user API token in Authorization: Bearer <token>. Tokens are accepted only when PYTORRENT_AUTH_ENABLE=1.",
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "description": "Optional Flask session cookie. Required only when PYTORRENT_AUTH_ENABLE=1 and no API token is used.",
        "in": "cookie",
        "name": "session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "pyTorrent API",
    "version": "0.0.1"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/automations": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutomationListResponse"
                }
              }
            },
            "description": "OK"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "List automation rules and history"
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutomationRule"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutomationListResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Create or update automation rule"
      }
    },
    "/api/automations/check": {
      "post": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutomationRunResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Run automation check immediately"
      }
    },
    "/api/automations/export": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutomationExportResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Export automation rules"
      }
    },
    "/api/automations/history": {
      "delete": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiOk"
                    },
                    {
                      "properties": {
                        "cleanup": {
                          "$ref": "#/components/schemas/CleanupSummary"
                        },
                        "deleted": {
                          "type": "integer"
                        },
                        "history": {
                          "items": {
                            "additionalProperties": true,
                            "type": "object"
                          },
                          "type": "array"
                        }
                      },
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Clear automation execution history"
      }
    },
    "/api/automations/import": {
      "post": {
        "parameters": [
          {
            "in": "query",
            "name": "replace",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutomationImportResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Import automation rules"
      }
    },
    "/api/automations/{rule_id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutomationListResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Delete automation rule"
      }
    },
    "/api/automations/{rule_id}/run": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutomationRunResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Run one automation rule immediately"
      }
    },
...
  }
}